+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.86.16
 
+- swift: Make `AciInfo` and `E164Info` constructors public